Split out `persistLeaves` test to a separate file
